% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.input.r
\name{deg2area_ha}
\alias{deg2area_ha}
\title{convert degree of latitue [deg] to area [Ha]}
\usage{
deg2area_ha(lat, res = 0.5)
}
\arguments{
\item{lat}{latitue}

\item{res}{resolution}
}
\value{
area in Ha
}
\description{
convert degree of latitue [deg] to area [Ha]
}

