% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/input.global.mean.r
\name{input.global.mean}
\alias{input.global.mean}
\title{Aggregate the global average of selected years and bands}
\usage{
input.global.mean(filename, years = NULL, bands = NULL)
}
\arguments{
\item{bands}{e.g. c(1:12)}

\item{year}{absolute values of years, e.g. c(1901:2000)}
}
\description{
Aggregate the global average of selected years and bands
}

