\name{LPJmL.RUtil-package}
\alias{LPJmL.RUtil-package}
\alias{LPJmL.RUtil}
\docType{package}
\title{
LPJmL Utilities
}
\description{
reading, writting, visualising and data manipulating utility functions related with LPJmL.
\details{
\tabular{ll}{
Package: \tab LPJmL.RUtil\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-03-23\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
S.SHI

Maintainer: Who to complain to <yourfault@somewhere.net>
S.SHI
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
