% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lpjncdf.r
\name{lpjoutput2ncdf}
\alias{lpjoutput2ncdf}
\title{convert any lpjoutput to ncdf}
\usage{
lpjoutput2ncdf(lpjoutput)
}
\arguments{
\item{lpjoutput:}{list}
}
\description{
convert any lpjoutput to ncdf
}

