% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map.r
\name{map.build}
\alias{map.build}
\title{Convert vector data to raster}
\usage{
map.build(raw_)
}
\arguments{
\item{raw_}{vector}
}
\value{
2-D array [NR, NC]
}
\description{
Convert vector data to raster
}

