% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lpjncdf.r
\name{new.var.ncdf}
\alias{new.var.ncdf}
\title{create an empty ncdf file with single variable.}
\usage{
new.var.ncdf(ncfile, lpjgrid_path, var_name, units, time_start, time_interval,
  time_dim, longname = var_name, missval = 1e+32)
}
\arguments{
\item{lpjgrid:}{the path of LPJ grid}

\item{var_name:}{variable name}

\item{time_start:}{start year for yearly output, start month for monthly output.
e.g. "1900" and "1900-01-01"}

\item{time_interval:}{"years" or "months" or "days"}

\item{time_dim:}{length of time dimention, e.g. months * years}

\item{lonname:}{the description of the variable}
}
\value{
list(cout, vardef): cout is the new ncdf file, and vardef \
is the definition of the all variables defined
}
\description{
create an empty ncdf file with single variable.
}

