% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/output.operation.r
\name{output.operation}
\alias{output.operation}
\title{output.operation
Do user defined operations on LPJ outputs.}
\usage{
output.operation(outfile, operation, newfile = paste(outfile[["path"]],
  "_new", sep = ""), VERBOSE = FALSE)
}
\arguments{
\item{outfile}{the information list of LPJ output, see example.}

\item{operation}{a user defined function/operation. Note: this
function should have only one scalar/vector as an argument
and return a scalar/vector value too.}
}
\description{
Convert one LPJ output file by a user defined operation.
             Write out new data in LPJ output (clm)
             format.
}
\examples{
lpjoutfile<-list()
lpjoutfile[["path"]]="../tests/test_data/soilc.bin"
lpjoutfile[["nyears"]]=40
lpjoutfile[["nbands"]]=1
lpjoutfile[["ncells"]]=67420
lpjoutfile[["start_year"]]=1901

fun<-function(old){ # here identify the user operation
    new <- old
    new[old>0] <- old/1000
    return((old/10000))
}
output.operation(lpjoutfile, fun, newfile="soilc_new.bin")
}

