% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.input.r
\name{read.input.grid}
\alias{read.input.grid}
\title{Read input grid (clm), return global values lon, lat, EAST, SOUTH, WEST, NORTH, RES, NC, int_lon, ind_lat ...}
\usage{
read.input.grid(path.in)
}
\arguments{
\item{path.in}{file location of grid.bin}
}
\value{
lon vector longitiute

lat vector latitide
}
\description{
Read input grid (clm), return global values lon, lat, EAST, SOUTH, WEST, NORTH, RES, NC, int_lon, ind_lat ...
}

