% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.input.r
\name{read.input.yearband}
\alias{read.input.yearband}
\title{Read one year and one band of LPJ clm data, and return a vector of the select year and band.}
\usage{
read.input.yearband(filename, year, band, data.size)
}
\arguments{
\item{filename}{input file path}

\item{year}{absolute value of select year, e.g. 1900}

\item{band}{band}

\item{data.size}{data size of input data, generally equal to 2.}
}
\value{
vector of npix
}
\description{
Read one year and one band of LPJ clm data, and return a vector of the select year and band.
}
\examples{
read.input.yearband("temp.clm", 1983, 1, 2)
}

