% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.input.r
\name{read.input.yearbands}
\alias{read.input.yearbands}
\title{Read one year and all bands of LPJ clm data, and return an ncells by nbands
array.}
\usage{
read.input.yearbands(filename, year, bands = NULL, data.size)
}
\arguments{
\item{filename}{input file path}

\item{year}{absolute value of select year, e.g. 1900}

\item{bands}{a vector of selected bands, if not specified read all bands.}

\item{data.size}{data size of input data, generally equal to 2.}
}
\value{
vector of $ncells \times bands$
}
\description{
Read one year and all bands of LPJ clm data, and return an ncells by nbands
array.
}
\examples{
read.input.yearband("temp.clm", 1983, 1, 2)
}

