% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.output.r
\name{read.output.grid}
\alias{read.output.grid}
\title{read output grid and return global variable lon, lat, and index.}
\usage{
read.output.grid(path, file, ncells)
}
\arguments{
\item{path}{output path.}

\item{ncells}{ncells.}
}
\description{
read output grid and return global variable lon, lat, and index.
}

