% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.output.r
\name{read.output.yearblock}
\alias{read.output.yearblock}
\title{read output by given a year block}
\usage{
read.output.yearblock(filename, syear, lyear, start_year, ncells, nyears,
  nbands, data.size = 4)
}
\arguments{
\item{syear}{start year of the year block}

\item{lyear}{end year of the year block}
}
\description{
read output by given a year block
}

