% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write.output.r
\name{write.output.yearband}
\alias{write.output.yearband}
\title{Write data of a selected band and year of LPJ binary output.}
\usage{
write.output.yearband(filename, data, year, band, start_year, ncells, nyears,
  nbands, data.size = 4)
}
\arguments{
\item{filename}{output file name}

\item{year}{select year}

\item{band}{select band}

\item{start_year}{start year of the LPJ output}

\item{ncells}{ncells of the LPJ output}

\item{nyears}{nyears of the LPJ output}

\item{nbands}{nbands of the LPJ output}

\item{data.size}{data size, which in general equal to 4.}
}
\value{
data in vector with ncells elements
}
\description{
Write data of a selected band and year of LPJ binary output.
}
\examples{
data<-read.output.yearband("mnpp.bin", year=1982, band=2,
                            nyears=1900, ncells=67420, nyears=120, nbands=12)
}

